\name{simDat4}
\alias{simDat4}
\docType{data}
\title{
Simulated data.
}
\description{
This dataset is used in several documentation files to illustrate the
different functionalities of the package.
}
\usage{data("simDat4")}
\format{
  A data frame with 300 observations on the following 9 variables. 
  \describe{
    \item{\code{Z}}{Treatment indicator.}
    \item{\code{Y}}{Observed outcome.}
    \item{\code{X1}}{Continuous covariates with a large proportion of
	zeros.}
    \item{\code{X2}}{A categorical variable with 2 levels (not expressed
      as factor).}
    \item{\code{X3}}{A categorical variable with 3 levels (not expressed
      as factor).}
    \item{\code{X4}}{A binary variable.}          
  }
}

\keyword{datasets}
